requirements.txt
# Backend Architecture / Render Server Dependencies
fastapi
uvicorn
pydantic
langgraph
langchain
langchain-groq
pymongo
xgboost
pandas
numpy
joblib
python-dotenv
scikit-learn
PyPDF2
